pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ose"
requires-python = ">= 3.12"

dependencies = [
    "Flask==2.3.3",
    "pandas",
    "openpyxl",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-Injector",
    "GitHub-Flask",
    "jsonschema",
    "daff",
    "whoosh",
    "networkx",
    "pydot",
    "py-horned-owl",
    "ontoutils",
    "flask-cors==3.0.10",
    "flask-executor",
    "aiohttp",
    "async-lru",
    "Flask-Migrate",
    "pyyaml",
    "dacite"

]

dynamic = ["version"]

[tool.setuptools.packages.find]
include = ["ose*"]  # ["*"] by default

[tool.setuptools.dynamic]
version = {attr = "ose.__version__"}

[dependency-groups]
dev = [
    "flake8>=5.0.4",
]

[project.scripts]
ose = "ose.cli_wrapper:main"
